Pin `backend.dockerfile` to inboard 0.10.4
br3ndonland/inboard@242ff7a

inboard is being updated to Uvicorn 0.15, which introduces some new
configuration options. The inboard updates are still a work in progress,
hence the alpha releases.

In addition to the Uvicorn version check added to `inboard.start` in
br3ndonland/inboard@242ff7a, this commit will pin the inboard version in
the whythawk/full-stack-fastapi-postgresql `backend.dockerfile` to the
latest stable release. Once the next version of inboard is stable, the
Docker image version in `backend.dockerfile`, and the Python version in
`pyproject.toml`, can be upgraded in tandem.